Reusable `Compressor` & `Decompressor` in `Stream`
## Overview

To enable the reuse of `Compressor` and `Decompressor` instances between
`CompressionStream` and `DecompressionStream` wrappers respectively,
this adds an overloaded constructor to each of the `Stream` children
which allows optionally providing an existing Zstandard context that
will survive any `IDisposable` and `IAsyncDisposable` resource cleanup.

This is particularly useful when needing `Stream` semantics in
conjunction with a `Compressor` or `Decompressor` instance that has a
dictionary loaded or otherwise has a custom set of parameters applied.